requirements.txt
pyyaml
ruamel.yaml
pytest
hypothesis==3.88.3
